pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcc-validate"
version = "0.1.0"
description = "Validator CLI for Model Context Certificates (MCC-STD-001)"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
readme = "README.md"
authors = [
    {name = "MCC Standard Authors"},
]
keywords = ["mcc", "ai", "medical-device", "validation", "certificate"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.0",
    "jsonschema>=4.20",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
crypto = [
    "cryptography>=42.0",
]
diff = [
    "deepdiff>=7.0",
]
reports = [
    "jinja2>=3.1",
]
web = [
    "flask>=3.0",
]
all = [
    "mcc-validate[crypto,diff,reports,web]",
]

[project.scripts]
mcc-validate = "mcc_validate.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcc_validate"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
